refacto: init hooks in its own method and rename Roomspool to server
